Kconfig: ili9341: Add CONFIG_LV_DISPLAY_USE_RST
On some devices, the reset pin isn't wired up. For example on the
M5Core2, it's connected to the AXP192 PMIC instead of the ESP32
directly.

This option can allow reset control to be skipped in these cases, so a
GPIO doesn't get assigned.

Only adds support in the ili9341 driver.